pyproject.toml
[tool.poetry]
name = "youth-learning-server"
version = "1.0.0"
description = "北邮青年大学习后端"
authors = ["zekin <wzekin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
django-excel = "^0.0.10"
pyexcel-xlsx = "^0.6.0"
django-simpleui = "^2021.3"
PyJWT = "^2.0.1"
sentry-sdk = "^0.20.3"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.10.1"
black = "^20.8b1"
isort = "^5.7.0"

[tool.isort]
profile = "django"
src_paths = ["oauth"]
skip_glob = "*/migrations/*.py"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
